Rural Health Development/requirements.txt
streamlit==1.29.0
Pillow==10.0.1  # For image handling